+++ src/equations.jl
@@ -1,15 +1,15 @@
 abstract type ElementID end
 
 struct KernelID <: ElementID
-    time_position::Int64
+    time_position::Int
     is_forward::Bool  # Forwrad for moving from nodes[1] to nodes[2]
-    nodes::Tuple{Int64, Int64}  # connected nodes by a kernel (gate)
+    nodes::Tuple{Int, Int}  # connected nodes by a kernel (gate)
 end
 
 struct IMID <: ElementID
-    time_position::Int64
+    time_position::Int
     is_forward::Bool # dirrection of the influende matrix
-    nodes::Tuple{Int64, Int64}  # connected nodes by an influence matrix
+    nodes::Tuple{Int, Int}  # connected nodes by an influence matrix
 end
 
 IMID(kernel_id::KernelID) = IMID(kernel_id.time_position, kernel_id.is_forward, kernel_id.nodes)
