+++ test/utils.jl
@@ -5,22 +5,22 @@ export QSim, apply_two_qubit_gate!, apply_one_qubit_gate!, get_one_qubit_dens, r
 using LinearAlgebra
 using Random
 
-@inline insert_zero(index::UInt64, mask::UInt64, flipped_mask::UInt64)::UInt64 = ((index & mask) << 1) | (index & flipped_mask)
+@inline insert_zero(index::UInt, mask::UInt, flipped_mask::UInt)::UInt = ((index & mask) << 1) | (index & flipped_mask)
 
-@inline get_mask(pos::UInt64)::UInt64 = typemax(UInt64) << (pos - 1)
+@inline get_mask(pos::UInt)::UInt = typemax(UInt) << (pos - 1)
 
-@inline flip_mask(mask::UInt64)::UInt64 = ~mask
+@inline flip_mask(mask::UInt)::UInt = ~mask
 
-@inline get_stride(pos::UInt64)::UInt64 = 1 << (pos - 1)
+@inline get_stride(pos::UInt)::UInt = 1 << (pos - 1)
 
 mutable struct QSim{T<:Real}
     state::Vector{Complex{T}}
-    qubits_number::UInt64
+    qubits_number::UInt
     function QSim(::Type{T}, qubits_number::Int) where {T<:Real}
         elems_number = 2^qubits_number
         state = zeros(Complex{T}, elems_number)
         state[1] = one(Complex{T})
-        new{T}(state, convert(UInt64, qubits_number))
+        new{T}(state, convert(UInt, qubits_number))
     end
 end
 
@@ -31,14 +31,14 @@ function apply_two_qubit_gate!(qsim::QSim{T}, gate::Matrix{Complex{T}}, pos1::In
     @assert pos1 >= 1
     @assert pos2 <= qsim.qubits_number
     @assert pos2 >= 1
-    apply_two_qubit_gate!(qsim, gate, convert(UInt64, pos1), convert(UInt64, pos2))
+    apply_two_qubit_gate!(qsim, gate, convert(UInt, pos1), convert(UInt, pos2))
 end
 
 function apply_two_qubit_gate!(
     qsim::QSim{T},
     gate::Matrix{Complex{T}},
-    pos2::UInt64,
-    pos1::UInt64,
+    pos2::UInt,
+    pos1::UInt,
 ) where {T<:Real}
     mask1 = get_mask(pos1)
     mask2 = get_mask(pos2)
@@ -50,7 +50,7 @@ function apply_two_qubit_gate!(
     stride2 = get_stride(pos2)
     size = 1 << (qsim.qubits_number - 2)
     state = qsim.state
-    @inbounds for i in zero(UInt64):UInt64(size - 1)
+    @inbounds for i in zero(UInt):UInt(size - 1)
         bi = insert_zero(insert_zero(i, max_mask, flipped_max_mask), min_mask, flipped_min_mask)
         tmp1 = (
               gate[begin, begin    ] * state[begin + bi]
@@ -86,16 +86,16 @@ function apply_one_qubit_gate!(qsim::QSim{T}, gate::Matrix{Complex{T}}, pos::Int
     @assert size(gate) == (2, 2)
     @assert pos <= qsim.qubits_number
     @assert pos >= 1
-    apply_one_qubit_gate!(qsim, gate, convert(UInt64, pos))
+    apply_one_qubit_gate!(qsim, gate, convert(UInt, pos))
 end
 
-function apply_one_qubit_gate!(qsim::QSim{T}, gate::Matrix{Complex{T}}, pos::UInt64) where {T<:Real}
+function apply_one_qubit_gate!(qsim::QSim{T}, gate::Matrix{Complex{T}}, pos::UInt) where {T<:Real}
     mask = get_mask(pos)
     flipped_mask = flip_mask(mask)
     stride = get_stride(pos)
     size = 1 << (qsim.qubits_number - 1)
     state = qsim.state
-    @inbounds for i in zero(UInt64):UInt64(size - 1)
+    @inbounds for i in zero(UInt):UInt(size - 1)
         bi = insert_zero(i, mask, flipped_mask)
         tmp = gate[begin, begin] * state[begin + bi] + gate[begin, begin + 1] * state[begin + bi + stride]
         state[begin + bi + stride] = gate[begin + 1, begin] * state[begin + bi] + gate[begin + 1, begin + 1] * state[begin + bi + stride]
@@ -106,17 +106,17 @@ end
 function get_one_qubit_dens(qsim::QSim{T}, pos::Int) where {T<:Real}
     @assert pos <= qsim.qubits_number
     @assert pos >= 1
-    get_one_qubit_dens(qsim, convert(UInt64, pos))
+    get_one_qubit_dens(qsim, convert(UInt, pos))
 end
 
-function get_one_qubit_dens(qsim::QSim{T}, pos::UInt64) where {T<:Real}
+function get_one_qubit_dens(qsim::QSim{T}, pos::UInt) where {T<:Real}
     dens = zeros(Complex{T}, 2, 2)
     mask = get_mask(pos)
     flipped_mask = flip_mask(mask)
     stride = get_stride(pos)
     size = 1 << (qsim.qubits_number - 1)
     state = qsim.state
-    @inbounds for i in zero(UInt64):UInt64(size - 1)
+    @inbounds for i in zero(UInt):UInt(size - 1)
         bi = insert_zero(i, mask, flipped_mask)
         dens[begin,     begin    ] += state[begin + bi         ] * conj(state[begin + bi])
         dens[begin,     begin + 1] += state[begin + bi         ] * conj(state[begin + bi + stride])
